Some clippy linting (some of which is temporary

+++ src/buffer.rs
@@ -50,7 +50,7 @@ impl Buffer {
 
     pub fn insert(&mut self, x: usize, y: usize, c: char) {
         if let Some(line) = self.lines.get_mut(y) {
-            (*line).insert(x as usize, c);
+            (*line).insert(x, c);
         }
     }
 
@@ -60,7 +60,7 @@ impl Buffer {
 
     pub fn remove(&mut self, x: usize, y: usize) {
         if let Some(line) = self.lines.get_mut(y) {
-            (*line).remove(x as usize);
+            (*line).remove(x);
         }
     }
